webapp/requirements.txt
azure-data-tables
flask
msal
azure-identity
azure-mgmt-web
azure-monitor-ingestion